Небольшой рефакторинг худовых подсказок
Теперь они не будут показываться, когда не должны, заодно добавил подсказки для таскания предметов (которое на самом деле работало и на оригинальном движке, если кто не знал, если визуал предмета, ящика например, прописан в [ph_capture_visuals], просто подсказок не было):
```xml
<string id="inventory_item_use_or_drag">
	<text>Поднять предмет ($$ACTION_USE$$) \n Перетащить предмет (SHIFT+$$ACTION_USE$$)</text>
</string>
<string id="game_object_drag">
	<text>Перетащить предмет (SHIFT+$$ACTION_USE$$)</text>
</string>
```
Вернул вызов use_callback туда, где он был раньше,
Поправил предупреждения и тп.

+++ ogsr_engine/xrGame/ui/UITradeWnd.cpp
@@ -549,7 +549,7 @@ void CUITradeWnd::UpdatePrices()
 	m_iOthersTradePrice = CalcItemsPrice	(&m_uidata->UIOthersTradeList,	m_pOthersTrade, false);
 
 	if ( !m_pOthersInvOwner->InfinitiveMoney() ) {
-          int others_money = (int)m_pOthersInvOwner->get_money();
+          u32 others_money = m_pOthersInvOwner->get_money();
           if ( others_zero_trade && m_iOurTradePrice > others_money )
             m_iOurTradePrice = others_money;
 	}
